Set safeTxGas for newly created transactions
