+++ src/logic/safe/transactions/gasNew.js
@@ -1,8 +1,10 @@
 // @flow
 import GnosisSafeSol from '@gnosis.pm/safe-contracts/build/contracts/GnosisSafe.json'
+import { BigNumber } from 'bignumber.js'
 
 import { CALL } from '.'
 
+import { getGnosisSafeInstanceAt } from '~/logic/contracts/safeContracts'
 import { generateSignaturesFromTxConfirmations } from '~/logic/safe/safeTxSigner'
 import { ZERO_ADDRESS } from '~/logic/wallets/ethAddresses'
 import { calculateGasOf, calculateGasPrice } from '~/logic/wallets/ethTransactions'
@@ -36,11 +38,33 @@ export const estimateTxGasCosts = async (
               '',
             )}000000000000000000000000000000000000000000000000000000000000000001`
       txData = await safeInstance.methods
-        .execTransaction(to, tx ? tx.value : 0, data, CALL, 0, 0, 0, ZERO_ADDRESS, ZERO_ADDRESS, signatures)
+        .execTransaction(
+          to,
+          tx ? tx.value : 0,
+          data,
+          CALL,
+          tx ? tx.safeTxGas : 0,
+          0,
+          0,
+          ZERO_ADDRESS,
+          ZERO_ADDRESS,
+          signatures,
+        )
         .encodeABI()
     } else {
       const txHash = await safeInstance.methods
-        .getTransactionHash(to, tx ? tx.value : 0, data, CALL, 0, 0, 0, ZERO_ADDRESS, ZERO_ADDRESS, nonce)
+        .getTransactionHash(
+          to,
+          tx ? tx.value : 0,
+          data,
+          CALL,
+          tx ? tx.safeTxGas : 0,
+          0,
+          0,
+          ZERO_ADDRESS,
+          ZERO_ADDRESS,
+          nonce,
+        )
         .call({
           from,
         })
@@ -56,5 +80,35 @@ export const estimateTxGasCosts = async (
     console.error(err)
 
     return 10000
+  }
+}
+
+export const estimateSafeTxGas = async (
+  safe: any,
+  safeAddress: string,
+  data: string,
+  to: string,
+  valueInWei: number | string,
+  operation: number,
+) => {
+  try {
+    let safeInstance = safe
+    if (!safeInstance) {
+      safeInstance = await getGnosisSafeInstanceAt(safeAddress)
+    }
+
+    const estimateData = safeInstance.contract.methods.requiredTxGas(to, valueInWei, data, operation).encodeABI()
+    const estimateResponse = await getWeb3().eth.call({
+      to: safeAddress,
+      from: safeAddress,
+      data: estimateData,
+    })
+    const txGasEstimate = new BigNumber(estimateResponse.substring(138), 16)
+
+    // Add 10k else we will fail in case of nested calls
+    return txGasEstimate.toNumber() + 10000
+  } catch (error) {
+    console.error('Error calculating tx gas estimation', error)
+    return 0
   }
 }
